+++ src/Service/Service.Test/TokenModules/AESEncryptionModuleTest.cs
@@ -24,7 +24,7 @@ namespace Service.Test.TokenModules
                 0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07,
                 0x08, 0x09 ,0x0A, 0x0B, 0x0C, 0x0D, 0x0E, 0x0F };
 
-        private byte[] keyType = BitConverter.GetBytes((uint)Pkcs11KeyType.AES).Reverse().ToArray();
+        private byte[] keyType = BitConverter.GetBytes((ulong)Pkcs11KeyType.AES).Reverse().ToArray();
